added unplaced industry component and constants file
