+++ Frontend/not-dominion/src/app/page.tsx
@@ -1,10 +1,7 @@
 import { StartGame } from './gameplay_module/State/Game'
-import IndustryTile from './components/IndustryTile'
 
 export default function Home() {
   const gameData = StartGame(4);
-  const red = "#6e0e08"
-  const grey = "#302c2c"
   return (
     <>
     </>
