+++ Frontend/not-dominion/src/app/utils/constants.ts
@@ -1,0 +1,3 @@
+export const grey = "#222426"
+
+export const red = "#6e0e08"
